refactor(seed): modularize seed files into separate functions
