+++ prisma/seeds/create-specialties.ts
@@ -1,0 +1,11 @@
+import { PrismaClient } from '../../generated/prisma';
+
+export const CreateSpecialties = async () => {
+	const prisma = new PrismaClient();
+
+	await prisma.specialty.createMany({
+		data: [{ name: 'Psiquiatria' }, { name: 'Psicologia' }],
+	});
+
+	await prisma.$disconnect();
+};
