+++ prisma/seeds/create-types-of-treatment.ts
@@ -1,0 +1,14 @@
+import { PrismaClient } from '../../generated/prisma';
+
+export const CreateTypesOfTreatment = async () => {
+	const prisma = new PrismaClient();
+
+	await prisma.typesOfTreatment.createMany({
+		data: [
+			{ name: 'Psiquiatria Infantil' },
+			{ name: 'Terapia de Casais' },
+			{ name: 'Terapia Individual' },
+		],
+	});
+	await prisma.$disconnect();
+};
